Exposing (publicly) necessary ModuleScopeBuilder methods as sealed interface
